backend/pyproject.toml
[project]
name = "sopher-ai-backend"
version = "0.1.0"
description = "AI book-writing system backend"
authors = [{name = "sopher.ai", email = "admin@sopher.ai"}]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "sse-starlette>=2.1.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "python-jose[cryptography]>=3.3.0",
    "prometheus-client>=0.20.0",
    "openai==1.68.2",
    "litellm==1.74.9",
    "crewai==0.80.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "cryptography>=42.0.0",
    "python-multipart>=0.0.9",
    "websockets>=12.0",
    "tenacity>=8.3.0",
    "alembic>=1.13.0",
    "passlib[bcrypt]>=1.7.4",
    "authlib>=1.3.0",
    "setuptools>=69.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "hypothesis>=6.100.0",
    "black>=24.4.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=69.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]